@@ -1,12 +1,12 @@
 name: pull_to_refresh_notification
 description: Flutter plugin for building pull to refresh effects with PullToRefreshNotification and PullToRefreshContainer quickly.
-version: 1.1.2
+version: 2.0.0
 homepage: https://github.com/fluttercandies/pull_to_refresh_notification
 
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
-  flutter: ">=1.12.13"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=2.0.0"
 
 dependencies:
   flutter:
