@@ -3,7 +3,7 @@
 // Auto generated by https://github.com/fluttercandies/ff_annotation_route
 // **************************************************************************
 
-import 'package:ff_annotation_route/ff_annotation_route.dart';
+import 'package:ff_annotation_route_library/ff_annotation_route_library.dart';
 import 'package:flutter/widgets.dart';
 import 'pages/complex/pull_to_refresh_candies.dart';
 import 'pages/complex/pull_to_refresh_chat.dart';
@@ -12,12 +12,18 @@ import 'pages/main_page.dart';
 import 'pages/simple/pull_to_refresh_appbar.dart';
 import 'pages/simple/pull_to_refresh_image.dart';
 
-RouteResult getRouteResult({String name, Map<String, dynamic> arguments}) {
-  arguments = arguments ?? const <String, dynamic>{};
+// ignore_for_file: prefer_const_literals_to_create_immutables,unused_local_variable
+FFRouteSettings getRouteSettings({
+  @required String name,
+  Map<String, dynamic> arguments,
+}) {
+  final Map<String, dynamic> safeArguments =
+      arguments ?? const <String, dynamic>{};
   switch (name) {
     case 'fluttercandies://PullToRefreshAppbar':
-      return RouteResult(
+      return FFRouteSettings(
         name: name,
+        arguments: arguments,
         widget: PullToRefreshAppbar(),
         routeName: 'PullToRefreshAppbar',
         description:
@@ -25,8 +31,9 @@ RouteResult getRouteResult({String name, Map<String, dynamic> arguments}) {
         exts: <String, dynamic>{'group': 'Simple', 'order': 0},
       );
     case 'fluttercandies://PullToRefreshCandies':
-      return RouteResult(
+      return FFRouteSettings(
         name: name,
+        arguments: arguments,
         widget: PullToRefreshCandies(),
         routeName: 'PullToRefreshCandies',
         description:
@@ -34,8 +41,9 @@ RouteResult getRouteResult({String name, Map<String, dynamic> arguments}) {
         exts: <String, dynamic>{'group': 'Complex', 'order': 1},
       );
     case 'fluttercandies://PullToRefreshChat':
-      return RouteResult(
+      return FFRouteSettings(
         name: name,
+        arguments: arguments,
         widget: PullToRefreshChat(),
         routeName: 'PullToRefreshChat',
         description:
@@ -43,8 +51,9 @@ RouteResult getRouteResult({String name, Map<String, dynamic> arguments}) {
         exts: <String, dynamic>{'group': 'Complex', 'order': 2},
       );
     case 'fluttercandies://PullToRefreshHeader':
-      return RouteResult(
+      return FFRouteSettings(
         name: name,
+        arguments: arguments,
         widget: PullToRefreshHeader(),
         routeName: 'PullToRefreshHeader',
         description:
@@ -52,8 +61,9 @@ RouteResult getRouteResult({String name, Map<String, dynamic> arguments}) {
         exts: <String, dynamic>{'group': 'Complex', 'order': 0},
       );
     case 'fluttercandies://PullToRefreshImage':
-      return RouteResult(
+      return FFRouteSettings(
         name: name,
+        arguments: arguments,
         widget: PullToRefreshImage(),
         routeName: 'PullToRefreshImage',
         description:
@@ -61,56 +71,23 @@ RouteResult getRouteResult({String name, Map<String, dynamic> arguments}) {
         exts: <String, dynamic>{'group': 'Simple', 'order': 1},
       );
     case 'fluttercandies://demogrouppage':
-      return RouteResult(
+      return FFRouteSettings(
         name: name,
+        arguments: arguments,
         widget: DemoGroupPage(
-          keyValue:
-              arguments['keyValue'] as MapEntry<String, List<DemoRouteResult>>,
+          keyValue: asT<MapEntry<String, List<DemoRouteResult>>>(
+              safeArguments['keyValue']),
         ),
         routeName: 'DemoGroupPage',
       );
     case 'fluttercandies://mainpage':
-      return RouteResult(
+      return FFRouteSettings(
         name: name,
+        arguments: arguments,
         widget: MainPage(),
         routeName: 'MainPage',
       );
     default:
-      return const RouteResult(name: 'flutterCandies://notfound');
+      return const FFRouteSettings(name: '404', routeName: '404_page');
   }
-}
-
-class RouteResult {
-  const RouteResult({
-    @required this.name,
-    this.widget,
-    this.showStatusBar = true,
-    this.routeName = '',
-    this.pageRouteType,
-    this.description = '',
-    this.exts,
-  });
-
-  /// The name of the route (e.g., "/settings").
-  ///
-  /// If null, the route is anonymous.
-  final String name;
-
-  /// The Widget return base on route
-  final Widget widget;
-
-  /// Whether show this route with status bar.
-  final bool showStatusBar;
-
-  /// The route name to track page
-  final String routeName;
-
-  /// The type of page route
-  final PageRouteType pageRouteType;
-
-  /// The description of route
-  final String description;
-
-  /// The extend arguments
-  final Map<String, dynamic> exts;
 }
