@@ -1,8 +1,7 @@
+import 'package:ff_annotation_route_library/ff_annotation_route_library.dart';
 import 'package:flutter/cupertino.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'example_route.dart';
-import 'example_route_helper.dart';
 import 'example_routes.dart';
 
 void main() => runApp(MyApp());
@@ -12,37 +11,29 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'create_package demo',
+      title: 'pull_to_refresh_notification demo',
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
       initialRoute: Routes.fluttercandiesMainpage,
       onGenerateRoute: (RouteSettings settings) {
-        //when refresh web, route will as following
-        //   /
-        //   /fluttercandies:
-        //   /fluttercandies:/
-        //   /fluttercandies://mainpage
-        if (kIsWeb && settings.name.startsWith('/')) {
-          return onGenerateRouteHelper(
-            settings.copyWith(name: settings.name.replaceFirst('/', '')),
-            notFoundFallback:
-                getRouteResult(name: Routes.fluttercandiesMainpage).widget,
-          );
-        }
-        return onGenerateRouteHelper(settings,
-            builder: (Widget child, RouteResult result) {
-          if (settings.name == Routes.fluttercandiesMainpage ||
-              settings.name == Routes.fluttercandiesDemogrouppage ||
-              settings.name == Routes.fluttercandiesPullToRefreshAppbar) {
-            return child;
-          }
-          return CommonWidget(
-            child: child,
-            result: result,
-          );
-        });
+        return onGenerateRoute(
+          settings: settings,
+          getRouteSettings: getRouteSettings,
+          routeSettingsWrapper: (FFRouteSettings ffRouteSettings) {
+            if (ffRouteSettings.name == Routes.fluttercandiesMainpage ||
+                ffRouteSettings.name == Routes.fluttercandiesDemogrouppage ||
+                settings.name == Routes.fluttercandiesPullToRefreshAppbar) {
+              return ffRouteSettings;
+            }
+            return ffRouteSettings.copyWith(
+                widget: CommonWidget(
+              child: ffRouteSettings.widget,
+              title: ffRouteSettings.routeName,
+            ));
+          },
+        );
       },
     );
   }
@@ -51,17 +42,17 @@ class MyApp extends StatelessWidget {
 class CommonWidget extends StatelessWidget {
   const CommonWidget({
     this.child,
-    this.result,
+    this.title,
   });
   final Widget child;
-  final RouteResult result;
+  final String title;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         title: Text(
-          result.routeName,
+          title,
         ),
       ),
       body: child,
